Moving elastic-http-frontend-connector to frontend_connectors package
